#849@minor: Adds support for triggering an error event when attempting to fetch a CSS file or Javascript file when it has been disabled using "window.happyDOM.settings.*".
